feat(ivy): generate pipe references and definitions
